Make folders start with capital letter, adjust namespaces, move model classes to separate files
