+++ Models/Config.cs
@@ -1,0 +1,7 @@
+﻿namespace OsuMultiplayerLobbyFinder.Models
+{
+    public class Config(string apiKey)
+    {
+        public string ApiKey { get; set; } = apiKey;
+    }
+}
