+++ Models/Event.cs
@@ -1,0 +1,10 @@
+﻿namespace OsuMultiplayerLobbyFinder.Models;
+
+public class Event
+{
+    public string display_html { get; set; }
+    public string beatmap_id { get; set; }
+    public string beatmapset_id { get; set; }
+    public string date { get; set; }
+    public string epicfactor { get; set; }
+}
