+++ Models/Lobby/Lobby.cs
@@ -1,0 +1,7 @@
+namespace OsuMultiplayerLobbyFinder.Models.Lobby;
+
+public class Lobby
+{
+    public Match match { get; set; }
+    public Game[] games { get; set; }
+}
