+++ Models/Lobby/Score.cs
@@ -1,0 +1,20 @@
+namespace OsuMultiplayerLobbyFinder.Models.Lobby;
+
+public class Score
+{
+    public string slot { get; set; }
+    public string team { get; set; }
+    public string user_id { get; set; }
+    public string score { get; set; }
+    public string maxcombo { get; set; }
+    public string rank { get; set; }
+    public string count50 { get; set; }
+    public string count100 { get; set; }
+    public string count300 { get; set; }
+    public string countmiss { get; set; }
+    public string countgeki { get; set; }
+    public string countkatu { get; set; }
+    public string perfect { get; set; }
+    public string pass { get; set; }
+    public string enabled_mods { get; set; }
+}
