--- feature/matcher/ILobbyMatcher.cs
@@ -1,7 +1,0 @@
-﻿namespace OsuMultiplayerLobbyFinder.feature.matcher
-{
-    public interface ILobbyMatcher
-    {
-        public bool MatchLobby(LobbyModel lobby, string? lobbyName, string? playerId);
-    }
-}
