--- models/ConfigModel.cs
@@ -1,12 +1,0 @@
-﻿namespace OsuMultiplayerLobbyFinder.models
-{
-    public class ConfigModel
-    {
-        public ConfigModel(string apiKey)
-        {
-            ApiKey = apiKey;
-        }
-
-        public string ApiKey { get; set; }
-    }
-}
